+++ src/commands/snippetGit.ts
@@ -1,18 +1,22 @@
 import * as vscode from 'vscode';
+import { temporarilyEnableAll } from '../snippets/fileDisabler.js';
+import { findAllGlobalSnippetFiles } from '../snippets/locateSnippets.js';
+import { getGlobalSnippetFilesDir } from '../utils/fsInfo.js';
+import { refreshAll } from './snippetFile.js';
 
 async function initSnippetGitCommands(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		vscode.commands.registerCommand('snippetstudio.github.sync', async () => {
 			const { snippetSync } = await import('../git/snippetSync.js');
-			await snippetSync(context);
-			vscode.commands.executeCommand('snippetstudio.refresh');
-			vscode.commands.executeCommand('snippetstudio.refreshLocations');
+			const globals = await findAllGlobalSnippetFiles(getGlobalSnippetFilesDir());
+			await temporarilyEnableAll(globals, () => snippetSync(context));
+			refreshAll();
 		}),
 		vscode.commands.registerCommand('snippetstudio.github.merge', async () => {
 			const { snippetMerge } = await import('../git/snippetMerge.js');
-			await snippetMerge(context);
-			vscode.commands.executeCommand('snippetstudio.refresh');
-			vscode.commands.executeCommand('snippetstudio.refreshLocations');
+			const globals = await findAllGlobalSnippetFiles(getGlobalSnippetFilesDir());
+			await temporarilyEnableAll(globals, () => snippetMerge(context));
+			refreshAll();
 		})
 	);
 }